refactor: member action 훅으로 분리
